+++ rollup.config.ts
@@ -69,7 +69,9 @@ const createConfig = ({ output, min = false }) => {
             'isLowEndDevice',
             'isLowEndExperience',
             'cumulativeLayoutShift',
-            'serviceWorkerStatus'
+            'serviceWorkerStatus',
+            'storageEstimateQuota',
+            'storageEstimateUsage'
           ]
         }
       },